@@ -33,7 +33,8 @@ def test_ds(store: zarr.storage.FSStore) -> zarr.storage.FSStore:
 
     ds = xr.open_dataset(store, engine="zarr", chunks={})
     assert ds.title == (
-        "Global Precipitation Climatatology Project (GPCP) " "Climate Data Record (CDR), Daily V1.3"
+        "Global Precipitation Climatatology Project (GPCP) "
+        "Climate Data Record (CDR), Daily V1.3"
     )
     # Making sure that the native chunking is different from the dynamic chunking
     assert ds.chunks["time"][0] == 1
@@ -44,7 +45,9 @@ def test_ds(store: zarr.storage.FSStore) -> zarr.storage.FSStore:
 recipe = (
     beam.Create(pattern.items())
     | OpenURLWithFSSpec()
-    | OpenWithXarray(file_type=pattern.file_type, xarray_open_kwargs={"decode_coords": "all"})
+    | OpenWithXarray(
+        file_type=pattern.file_type, xarray_open_kwargs={"decode_coords": "all"}
+    )
     | StoreToZarr(
         store_name="gpcp.zarr",
         combine_dims=pattern.combine_dim_keys,
