@@ -42,7 +42,9 @@ def concat_merge_pattern():
     return make_concat_merge_pattern()
 
 
-@pytest.fixture(params=[dict(fsspec_open_kwargs={"block_size": "foo"}), dict(file_type="opendap")])
+@pytest.fixture(
+    params=[dict(fsspec_open_kwargs={"block_size": "foo"}), dict(file_type="opendap")]
+)
 def concat_merge_pattern_with_kwargs(request):
     return make_concat_merge_pattern(**request.param)
 
@@ -86,8 +88,9 @@ def test_pattern_from_file_sequence():
 
 
 @pytest.mark.parametrize("pickle", [False, True])
-def test_file_pattern_concat_merge(runtime_secrets, pickle, concat_merge_pattern_with_kwargs):
-
+def test_file_pattern_concat_merge(
+    runtime_secrets, pickle, concat_merge_pattern_with_kwargs
+):
     fp, times, varnames, format_function, kwargs = concat_merge_pattern_with_kwargs
 
     if runtime_secrets:
@@ -155,7 +158,6 @@ def test_incompatible_kwargs():
 
 @pytest.mark.parametrize("nkeep", [1, 2])
 def test_prune(nkeep, concat_merge_pattern_with_kwargs, runtime_secrets):
-
     fp = concat_merge_pattern_with_kwargs[0]
 
     if runtime_secrets:
@@ -188,16 +190,19 @@ def test_prune(nkeep, concat_merge_pattern_with_kwargs, runtime_secrets):
     assert get_kwargs(fp) == get_kwargs(fp_pruned)
 
 
-@pytest.mark.parametrize("file_type_value", [ft.value for ft in list(FileType)] + ["unsupported"])
+@pytest.mark.parametrize(
+    "file_type_value", [ft.value for ft in list(FileType)] + ["unsupported"]
+)
 def test_setting_file_types(file_type_value):
-
     file_type_kwargs = {"file_type": file_type_value}
 
     if not file_type_value == "unsupported":
         fp = make_concat_merge_pattern(**file_type_kwargs)[0]
         assert fp.file_type == FileType(file_type_value)
     else:
-        with pytest.raises(ValueError, match=rf"'{file_type_value}' is not a valid FileType"):
+        with pytest.raises(
+            ValueError, match=rf"'{file_type_value}' is not a valid FileType"
+        ):
             fp = make_concat_merge_pattern(**file_type_kwargs)[0]
 
 
